requirements.txt
pyserial
pycrc